A News Website Sample
Html, Css and Javascript are used for this website. There is a ''Click To Sign' button for login and subscripe. Slides are opening when clicked on it. They can be changed by prev and next buttons. At the footer there are  counters. They starts to count when website is opened.